add ability for KnownErrorCode which then frees up future error codes to be put into the DTO and used by code that knows about it(or for apps that have their own and can use server and client and we don't care about it or know about it).

+++ http/http-common/src/main/java/org/webpieces/httpcommon/impl/Http2EngineImpl.java
@@ -129,7 +129,7 @@ public abstract class Http2EngineImpl implements Http2Engine {
     // going to record what they told us.
     // these don't have to be concurrent-safe because the datalistener is virtually single threaded.
     boolean remoteGoneAway = false;
-    int goneAwayLastStreamId;
+    long goneAwayLastStreamId;
     Http2ErrorCode goneAwayErrorCode;
     DataWrapper additionalDebugData;
